Refactored integration tests to be based on the unittest module
Yes, the test_tick_rate is still failing.
